backend/API/requirement.txt
uwsgi==2.0.18
flask==1.1.1
Flask-RESTful==0.3.7
flask-cors==3.0.8
flask_upload==0.2.1
pymongo==3.10.1
mysql-connector-python==8.0.19
Flask-JWT-Extended==3.24.1
google-cloud-storage==1.26.0
